+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#224](https://github.com/mongoid/mongoid-slug/pull/224): Use Danger, PR linter - [@dblock](https://github.com/dblock).
 * [#222](https://github.com/mongoid/mongoid-slug/pull/225): Fix: `Mongo::Error::OperationFailure: E11000 duplicate key error index` error with blank slugs, default `_slugs` to `nil` instead of `[]` - [@dblock](https://github.com/dblock).
 * [#172](https://github.com/mongoid/mongoid-slug/pull/172): Improved handling of unique and sparse index constraints - [@johnnyshields](https://github.com/johnnyshields).
+* [#229](https://github.com/mongoid/mongoid-slug/pull/229): Upgraded to RuboCop 0.42.0 - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ## 5.2.0 (2016/01/03)